rename test mods
