+++ agent-control/tests/on_host/mod.rs
@@ -5,8 +5,8 @@ mod config_repository;
 mod consts;
 mod id;
 mod logging;
-mod oci_package_management;
-mod oci_registry;
+mod oci_artifact_downloader;
+mod oci_package_manager;
 mod opamp_auth;
 mod scenarios;
 pub mod tools;
