+++ agent-control/tests/on_host/oci_artifact_downloader.rs
